+++ queen/fitness_test.go
@@ -7,7 +7,7 @@ func TestPhenotype_CalcFitness(t *testing.T) {
 		pTest := phenotype{
 			//board: []int{7,5,4,3,1,6,0,2},
 			//board:[]int{6,4,2,7,5,3,1,0},
-			board: []int{2, 7, 3, 6, 0, 5, 1, 4},
+			board: []float64{2, 7, 3, 6, 0, 5, 1, 4},
 		}
 		pTest.Print()
 		//pTest := NewRandOneMaxGenotype().GetPhenotype()
